requirements.txt
--index-url https://download.pytorch.org/whl/cu118

torch
torchvision
torchaudio

transformers
matplotlib
seaborn
pillow
ipython
shapiq
notebook
